Move modules towards a 'configobj' package
* handles #72
* handles #31
* handles the core of #32
* using ImportWarning as appropriate for an import change
* tox passing for py27, i.e. structurally sound
* the shim should probably released one final time as
  'validate' to PyPI
