--- _version.py
@@ -1,1 +1,0 @@
-__version__ = '5.0.6'
