+++ setup.cfg
@@ -1,0 +1,17 @@
+#
+# Configuration for setuptools
+#
+
+[egg_info]
+tag_build = .dev0
+tag_date = false
+
+
+[sdist]
+formats = zip
+
+
+[bdist_wheel]
+# If you set this to 1, make sure you have a proper Travis CI build matrix,
+# and that your Trove classifiers state you support Python 2 and 3
+universal = 1
