+++ src/configobj/_version.py
@@ -1,0 +1,1 @@
+__version__ = '5.1.0'
