+++ tox.ini
@@ -1,12 +1,27 @@
 # content of: tox.ini, put in same dir as setup.py
 [tox]
-envlist = py26,py27,py32,py33
+envlist = py26,py27,py32,py33,py34
+
+
 [testenv]
-deps=pytest
+deps =
+    six
+    pytest
     pytest-cov
 setenv =
     PYTHONWARNINGS = always
-commands=python test_configobj.py
-     python validate.py
-     coverage run --source=configobj.py,validate.py -m py.test tests
-     coverage report
+commands =
+    python test_configobj.py
+    python -m configobj.validate
+    coverage run --source=configobj.py,validate.py -m py.test tests
+    coverage report
+
+
+[testenv:flake8]
+deps =
+    flake8==2.3.0
+    pep8==1.6.2
+
+; for now just informational (--exit-zero)
+commands =
+    flake8 --count --statistics --exit-zero src/configobj
